feat(ehr): moved into internal folder
Ref: #1040
