+++ packages/api/src/routes/jwt-token/internal/ehr/elation/jwt-token.ts
@@ -1,21 +1,21 @@
 import {
-  canvasDashJwtTokenDataSchema,
-  canvasDashSource,
-  canvasWebhookJwtTokenDataSchema,
-  canvasWebhookSource,
-} from "@metriport/shared/interface/external/ehr/canvas/jwt-token";
+  elationDashJwtTokenDataSchema,
+  elationDashSource,
+  elationWebhookJwtTokenDataSchema,
+  elationWebhookSource,
+} from "@metriport/shared/interface/external/ehr/elation/jwt-token";
 import { Request, Response } from "express";
 import Router from "express-promise-router";
 import httpStatus from "http-status";
 import z from "zod";
-import { checkJwtToken, saveJwtToken } from "../../../../external/ehr/jwt-token";
-import { requestLogger } from "../../../helpers/request-logger";
-import { asyncHandler, getAuthorizationToken } from "../../../util";
+import { checkJwtToken, saveJwtToken } from "../../../../../external/ehr/jwt-token";
+import { requestLogger } from "../../../../helpers/request-logger";
+import { asyncHandler, getAuthorizationToken } from "../../../../util";
 
 const router = Router();
 
 /**
- * GET /internal/token/canvas
+ * GET /internal/token/elation
  */
 router.get(
   "/",
@@ -24,7 +24,7 @@ router.get(
     const token = getAuthorizationToken(req);
     const tokenStatus = await checkJwtToken({
       token,
-      source: canvasDashSource,
+      source: elationDashSource,
     });
     return res.status(httpStatus.OK).json(tokenStatus);
   })
@@ -32,11 +32,11 @@ router.get(
 
 const createJwtSchema = z.object({
   exp: z.number(),
-  data: canvasDashJwtTokenDataSchema,
+  data: elationDashJwtTokenDataSchema,
 });
 
 /**
- * POST /internal/token/canvas
+ * POST /internal/token/elation
  */
 router.post(
   "/",
@@ -46,7 +46,7 @@ router.post(
     const data = createJwtSchema.parse(req.body);
     await saveJwtToken({
       token,
-      source: canvasDashSource,
+      source: elationDashSource,
       ...data,
     });
     return res.sendStatus(httpStatus.OK);
@@ -54,7 +54,7 @@ router.post(
 );
 
 /**
- * GET /internal/token/canvas/webhook
+ * GET /internal/token/elation/webhook
  */
 router.get(
   "/webhook",
@@ -63,7 +63,7 @@ router.get(
     const token = getAuthorizationToken(req);
     const tokenStatus = await checkJwtToken({
       token,
-      source: canvasWebhookSource,
+      source: elationWebhookSource,
     });
     return res.status(httpStatus.OK).json(tokenStatus);
   })
@@ -71,11 +71,11 @@ router.get(
 
 const createWebhookJwtSchema = z.object({
   exp: z.number(),
-  data: canvasWebhookJwtTokenDataSchema,
+  data: elationWebhookJwtTokenDataSchema,
 });
 
 /**
- * POST /internal/token/canvas/webhook
+ * POST /internal/token/elation/webhook
  */
 router.post(
   "/webhook",
@@ -85,7 +85,7 @@ router.post(
     const data = createWebhookJwtSchema.parse(req.body);
     await saveJwtToken({
       token,
-      source: canvasWebhookSource,
+      source: elationWebhookSource,
       ...data,
     });
     return res.sendStatus(httpStatus.OK);
